requirements.txt
appdirs==1.4.3
beautifulsoup4==4.8.1
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
cssselect==1.1.0
Django==2.2.8
django-braces==1.13.0
django-debug-toolbar==2.1
django-model-utils==4.0.0
django-star-ratings==0.9.1
fake-useragent==0.1.11
idna==2.8
lxml==4.4.2
parse==1.14.0
Pillow==6.2.1
psycopg2==2.8.4
pyee==6.0.0
pyppeteer==0.0.25
pyquery==1.4.1
python-decouple==3.3
python-slugify==4.0.0
pytz==2019.3
requests==2.22.0
requests-html==0.10.0
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
swapper==1.1.1
text-unidecode==1.3
tqdm==4.41.0
urllib3==1.25.7
w3lib==1.21.0
websockets==8.1
